+++ header-pwa.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * The Header for our theme in PWA.
+ *
+ * @package OceanWP WordPress theme
+ * @since   1.8.8
+ */
+
+?>
+<!DOCTYPE html>
+<html class="<?php echo esc_attr( oceanwp_html_classes() ); ?>" <?php language_attributes(); ?>>
+<head>
+	<meta charset="<?php bloginfo( 'charset' ); ?>">
+	<link rel="profile" href="https://gmpg.org/xfn/11">
+
+	<?php wp_head(); ?>
+</head>
+
+<body <?php body_class(); ?> <?php oceanwp_schema_markup( 'html' ); ?>>