+++ header.php
@@ -6,7 +6,6 @@
  */
 
 ?>
-
 <!DOCTYPE html>
 <html class="<?php echo esc_attr( oceanwp_html_classes() ); ?>" <?php language_attributes(); ?>>
 <head>