+++ inc/oceanwp-strings.php
@@ -15,7 +15,7 @@ if ( ! function_exists( 'oceanwp_theme_strings' ) ) {
 	/**
 	 * OceanWP Theme Strings
 	 *
-	 *  @author Amit Singh (@apprimit)
+	 *  @author Amit Singh (apprimit@gmail.com)
 	 *  @since 1.8.5
 	 *
 	 * @param  string  $value  String key.
@@ -58,6 +58,7 @@ if ( ! function_exists( 'oceanwp_theme_strings' ) ) {
 				'owp-string-woo-floating-bar-select-btn' => apply_filters( 'ocean_woo_floating_bar_select_btn', __( 'Select Options', 'oceanwp' ) ),
 				'owp-string-woo-floating-bar-selected'   => apply_filters( 'ocean_woo_floating_bar_selected', __( 'Selected:', 'oceanwp' ) ),
 				'owp-string-woo-floating-bar-out-stock'  => apply_filters( 'ocean_woo_floating_bar_out_stock', __( 'Out of stock', 'oceanwp' ) ),
+<<<<<<< HEAD
 				'owp-string-woo-nav-next-product'        => apply_filters( 'ocean_woo_nav_next_text', __( 'Next Product', 'oceanwp' ) ),
 				'owp-string-woo-nav-prev-product'        => apply_filters( 'ocean_woo_nav_prev_text', __( 'Previous Product', 'oceanwp' ) ),
 
@@ -66,6 +67,8 @@ if ( ! function_exists( 'oceanwp_theme_strings' ) ) {
 				'owp-string-mobile-search'               => apply_filters( 'ocean_wai_mobile_search', __( 'Search for:', 'oceanwp' ) ),
 				'owp-string-mobile-submit-search'        => apply_filters( 'ocean_wai_mobile_search_submit', __( 'Submit search', 'oceanwp' ) ),
 
+=======
+>>>>>>> dev-2
 			)
 		);