+++ inc/plugins/tgm-plugin-activation.php
@@ -7,7 +7,7 @@
 
 function oceanwp_tgmpa_register() {
 
-	// Get array of recommended plugins
+	// Get array of recommended plugins.
 	$plugins = array(
 
 		array(
@@ -16,25 +16,7 @@ function oceanwp_tgmpa_register() {
 			'required'			=> false,
 			'force_activation'	=> false,
 		),
-
-		array(
-			'name'				=> 'Elementor',
-			'slug'				=> 'elementor',
-			'required'			=> false,
-			'force_activation'	=> false,
-		),
-
 	);
-
-	// If WPForms Pro is not active, recommend WPForms
-	if ( ! class_exists( 'WPForms_Pro' ) ) {
-		$plugins[] = array(
-			'name'				=> 'WPForms',
-			'slug'				=> 'wpforms-lite',
-			'required'			=> false,
-			'force_activation'	=> false,
-		);
-	}
 
 	// Register notice
 	tgmpa( $plugins, array(